get workdir

+++ app/resources/jobs/jobs.py
@@ -174,7 +174,7 @@ class Job_queue(Resource):
 
 			print "WRKDIR", processes_wrkdir[counter]
 
-			if processes_wrkdir[counter] != "false":
+			if processes_wrkdir[counter] != "false" && processes_to_run[counter] == "true":
 				workdirPath = os.path.join(current_user.homedir, "jobs", args.project_id + "-" + args.pipeline_id, "work", processes_wrkdir[counter])
 				shutil.rmtree(workdirPath)
 				print workdirPath
